Separate out getMonthNames
